refactor config flow with helpers
